requirement.txt
apache-airflow==2.7.3  # Adjust based on your Airflow version
boto3==1.28.57         # AWS SDK for Python
numpy==1.26.2          # Latest stable version
pandas==2.1.3          # Latest stable version
psycopg2-binary==2.9.9 # PostgreSQL adapter for Python
